+++ config.go
@@ -140,6 +140,7 @@ type config struct {
 
 	MaxSrcDimension  int
 	MaxSrcResolution int
+	MaxSrcFileSize   int
 	MaxGifFrames     int
 
 	JpegProgressive bool
@@ -193,30 +194,37 @@ type config struct {
 	SentryDSN         string
 	SentryEnvironment string
 	SentryRelease     string
+
+	FreeMemoryInterval             int
+	DownloadBufferSize             int
+	GZipBufferSize                 int
+	BufferPoolCalibrationThreshold int
 }
 
 var conf = config{
-	Bind:                  ":8080",
-	ReadTimeout:           10,
-	WriteTimeout:          10,
-	DownloadTimeout:       5,
-	Concurrency:           runtime.NumCPU() * 2,
-	TTL:                   3600,
-	IgnoreSslVerification: false,
-	MaxSrcResolution:      16800000,
-	MaxGifFrames:          1,
-	AllowInsecure:         false,
-	SignatureSize:         32,
-	Quality:               80,
-	GZipCompression:       5,
-	UserAgent:             fmt.Sprintf("imgproxy/%s", version),
-	ETagEnabled:           false,
-	S3Enabled:             false,
-	WatermarkOpacity:      1,
-	BugsnagStage:          "production",
-	HoneybadgerEnv:        "production",
-	SentryEnvironment:     "production",
-	SentryRelease:         fmt.Sprintf("imgproxy/%s", version),
+	Bind:                           ":8080",
+	ReadTimeout:                    10,
+	WriteTimeout:                   10,
+	DownloadTimeout:                5,
+	Concurrency:                    runtime.NumCPU() * 2,
+	TTL:                            3600,
+	IgnoreSslVerification:          false,
+	MaxSrcResolution:               16800000,
+	MaxGifFrames:                   1,
+	AllowInsecure:                  false,
+	SignatureSize:                  32,
+	Quality:                        80,
+	GZipCompression:                5,
+	UserAgent:                      fmt.Sprintf("imgproxy/%s", version),
+	ETagEnabled:                    false,
+	S3Enabled:                      false,
+	WatermarkOpacity:               1,
+	BugsnagStage:                   "production",
+	HoneybadgerEnv:                 "production",
+	SentryEnvironment:              "production",
+	SentryRelease:                  fmt.Sprintf("imgproxy/%s", version),
+	FreeMemoryInterval:             10,
+	BufferPoolCalibrationThreshold: 1024,
 }
 
 func init() {
@@ -248,6 +256,7 @@ func init() {
 
 	intEnvConfig(&conf.MaxSrcDimension, "IMGPROXY_MAX_SRC_DIMENSION")
 	megaIntEnvConfig(&conf.MaxSrcResolution, "IMGPROXY_MAX_SRC_RESOLUTION")
+	intEnvConfig(&conf.MaxSrcFileSize, "IMGPROXY_MAX_SRC_FILE_SIZE")
 	intEnvConfig(&conf.MaxGifFrames, "IMGPROXY_MAX_GIF_FRAMES")
 
 	boolEnvConfig(&conf.JpegProgressive, "IMGPROXY_JPEG_PROGRESSIVE")
@@ -308,6 +317,11 @@ func init() {
 	strEnvConfig(&conf.SentryEnvironment, "IMGPROXY_SENTRY_ENVIRONMENT")
 	strEnvConfig(&conf.SentryRelease, "IMGPROXY_SENTRY_RELEASE")
 
+	intEnvConfig(&conf.FreeMemoryInterval, "IMGPROXY_FREE_MEMORY_INTERVAL")
+	intEnvConfig(&conf.DownloadBufferSize, "IMGPROXY_DOWNLOAD_BUFFER_SIZE")
+	intEnvConfig(&conf.GZipBufferSize, "IMGPROXY_GZIP_BUFFER_SIZE")
+	intEnvConfig(&conf.BufferPoolCalibrationThreshold, "IMGPROXY_BUFFER_POOL_CALIBRATION_THRESHOLD")
+
 	if len(conf.Keys) != len(conf.Salts) {
 		logFatal("Number of keys and number of salts should be equal. Keys: %d, salts: %d", len(conf.Keys), len(conf.Salts))
 	}
@@ -362,6 +376,10 @@ func init() {
 		logFatal("Max src resolution should be greater than 0, now - %d\n", conf.MaxSrcResolution)
 	}
 
+	if conf.MaxSrcFileSize < 0 {
+		logFatal("Max src file size should be greater than or equal to 0, now - %d\n", conf.MaxSrcFileSize)
+	}
+
 	if conf.MaxGifFrames <= 0 {
 		logFatal("Max GIF frames should be greater than 0, now - %d\n", conf.MaxGifFrames)
 	}
@@ -373,7 +391,7 @@ func init() {
 	}
 
 	if conf.GZipCompression < 0 {
-		logFatal("GZip compression should be greater than or quual to 0, now - %d\n", conf.GZipCompression)
+		logFatal("GZip compression should be greater than or equal to 0, now - %d\n", conf.GZipCompression)
 	} else if conf.GZipCompression > 9 {
 		logFatal("GZip compression can't be greater than 9, now - %d\n", conf.GZipCompression)
 	}
@@ -410,9 +428,29 @@ func init() {
 		logFatal("Can't use the same binding for the main server and Prometheus")
 	}
 
-	initDownloading()
+	if conf.FreeMemoryInterval <= 0 {
+		logFatal("Free memory interval should be greater than zero")
+	}
+
+	if conf.DownloadBufferSize < 0 {
+		logFatal("Download buffer size should be greater than or equal to 0")
+	} else if conf.DownloadBufferSize > int(^uint32(0)) {
+		logFatal("Download buffer size can't be greater than %d", ^uint32(0))
+	}
+
+	if conf.GZipBufferSize < 0 {
+		logFatal("GZip buffer size should be greater than or equal to 0")
+	} else if conf.GZipBufferSize > int(^uint32(0)) {
+		logFatal("GZip buffer size can't be greater than %d", ^uint32(0))
+	}
+
+	if conf.BufferPoolCalibrationThreshold < 64 {
+		logFatal("Buffer pool calibration threshold should be greater than or equal to 64")
+	}
+
 	initNewrelic()
 	initPrometheus()
+	initDownloading()
 	initErrorsReporting()
 	initVips()
 }