+++ download.go
@@ -8,8 +8,8 @@ import (
 	"image"
 	"io"
 	"io/ioutil"
+	"net"
 	"net/http"
-	"sync"
 	"time"
 
 	_ "image/gif"
@@ -25,21 +25,42 @@ var (
 	imageDataCtxKey = ctxKey("imageData")
 
 	errSourceDimensionsTooBig      = newError(422, "Source image dimensions are too big", "Invalid source image")
-	errSourceResolutionTooBig      = newError(422, "Source image resolution are too big", "Invalid source image")
+	errSourceResolutionTooBig      = newError(422, "Source image resolution is too big", "Invalid source image")
+	errSourceFileTooBig            = newError(422, "Source image file is too big", "Invalid source image")
 	errSourceImageTypeNotSupported = newError(422, "Source image type not supported", "Invalid source image")
 )
 
 const msgSourceImageIsUnreachable = "Source image is unreachable"
 
-var downloadBufPool = sync.Pool{
-	New: func() interface{} {
-		return new(bytes.Buffer)
-	},
+var downloadBufPool *bufPool
+
+type limitReader struct {
+	r    io.ReadCloser
+	left int
+}
+
+func (lr *limitReader) Read(p []byte) (n int, err error) {
+	n, err = lr.r.Read(p)
+	lr.left = lr.left - n
+
+	if err == nil && lr.left < 0 {
+		err = errSourceFileTooBig
+	}
+
+	return
+}
+
+func (lr *limitReader) Close() error {
+	return lr.r.Close()
 }
 
 func initDownloading() {
 	transport := &http.Transport{
-		Proxy: http.ProxyFromEnvironment,
+		Proxy:               http.ProxyFromEnvironment,
+		MaxIdleConns:        conf.Concurrency,
+		MaxIdleConnsPerHost: conf.Concurrency,
+		DisableCompression:  true,
+		Dial:                (&net.Dialer{KeepAlive: 600 * time.Second}).Dial,
 	}
 
 	if conf.IgnoreSslVerification {
@@ -47,7 +68,7 @@ func initDownloading() {
 	}
 
 	if conf.LocalFileSystemRoot != "" {
-		transport.RegisterProtocol("local", http.NewFileTransport(http.Dir(conf.LocalFileSystemRoot)))
+		transport.RegisterProtocol("local", newFsTransport())
 	}
 
 	if conf.S3Enabled {
@@ -62,6 +83,8 @@ func initDownloading() {
 		Timeout:   time.Duration(conf.DownloadTimeout) * time.Second,
 		Transport: transport,
 	}
+
+	downloadBufPool = newBufPool("download", conf.Concurrency, conf.DownloadBufferSize)
 }
 
 func checkDimensions(width, height int) error {
@@ -78,8 +101,11 @@ func checkDimensions(width, height int) error {
 
 func checkTypeAndDimensions(r io.Reader) (imageType, error) {
 	imgconf, imgtypeStr, err := image.DecodeConfig(r)
-	if err != nil {
+	if err == image.ErrFormat {
 		return imageTypeUnknown, errSourceImageTypeNotSupported
+	}
+	if err != nil {
+		return imageTypeUnknown, err
 	}
 
 	imgtype, imgtypeOk := imageTypes[imgtypeStr]
@@ -95,18 +121,29 @@ func checkTypeAndDimensions(r io.Reader) (imageType, error) {
 }
 
 func readAndCheckImage(ctx context.Context, res *http.Response) (context.Context, context.CancelFunc, error) {
-	buf := downloadBufPool.Get().(*bytes.Buffer)
+	var contentLength int
+
+	if res.ContentLength > 0 {
+		contentLength = int(res.ContentLength)
+	}
+
+	buf := downloadBufPool.Get(contentLength)
 	cancel := func() {
-		buf.Reset()
 		downloadBufPool.Put(buf)
 	}
 
-	imgtype, err := checkTypeAndDimensions(io.TeeReader(res.Body, buf))
+	body := res.Body
+
+	if conf.MaxSrcFileSize > 0 {
+		body = &limitReader{r: body, left: conf.MaxSrcFileSize}
+	}
+
+	imgtype, err := checkTypeAndDimensions(io.TeeReader(body, buf))
 	if err != nil {
 		return ctx, cancel, err
 	}
 
-	if _, err = buf.ReadFrom(res.Body); err != nil {
+	if _, err = buf.ReadFrom(body); err != nil {
 		return ctx, cancel, newError(404, err.Error(), msgSourceImageIsUnreachable)
 	}
 
@@ -136,10 +173,12 @@ func downloadImage(ctx context.Context) (context.Context, context.CancelFunc, er
 	req.Header.Set("User-Agent", conf.UserAgent)
 
 	res, err := downloadClient.Do(req)
+	if res != nil {
+		defer res.Body.Close()
+	}
 	if err != nil {
 		return ctx, func() {}, newError(404, err.Error(), msgSourceImageIsUnreachable)
 	}
-	defer res.Body.Close()
 
 	if res.StatusCode != 200 {
 		body, _ := ioutil.ReadAll(res.Body)